requirements.txt
PySide6>=6.8.0
PySide6-Addons>=6.8.0
appdirs>=1.4.4
mss>=9.0.0
pyobjc-framework-Quartz>=10.0; sys_platform == "darwin"
pynput>=1.7.7; sys_platform == "darwin"
